% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_create_business_report_schedule}
\alias{alexaforbusiness_create_business_report_schedule}
\title{Creates a recurring schedule for usage reports to deliver to the
specified S3 location with a specified daily or weekly interval}
\usage{
alexaforbusiness_create_business_report_schedule(ScheduleName,
  S3BucketName, S3KeyPrefix, Format, ContentRange, Recurrence,
  ClientRequestToken)
}
\arguments{
\item{ScheduleName}{The name identifier of the schedule.}

\item{S3BucketName}{The S3 bucket name of the output reports. If this isn't specified, the
report can be retrieved from a download link by calling
ListBusinessReportSchedule.}

\item{S3KeyPrefix}{The S3 key where the report is delivered.}

\item{Format}{[required] The format of the generated report (individual CSV files or zipped files
of individual files).}

\item{ContentRange}{[required] The content range of the reports.}

\item{Recurrence}{The recurrence of the reports. If this isn't specified, the report will
only be delivered one time when the API is called.}

\item{ClientRequestToken}{The client request token.}
}
\description{
Creates a recurring schedule for usage reports to deliver to the
specified S3 location with a specified daily or weekly interval.
}
\section{Request syntax}{
\preformatted{svc$create_business_report_schedule(
  ScheduleName = "string",
  S3BucketName = "string",
  S3KeyPrefix = "string",
  Format = "CSV"|"CSV_ZIP",
  ContentRange = list(
    Interval = "ONE_DAY"|"ONE_WEEK"
  ),
  Recurrence = list(
    StartDate = "string"
  ),
  ClientRequestToken = "string"
)
}
}

\keyword{internal}
