% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_delete_contact}
\alias{alexaforbusiness_delete_contact}
\title{Deletes a contact by the contact ARN}
\usage{
alexaforbusiness_delete_contact(ContactArn)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to delete.}
}
\description{
Deletes a contact by the contact ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_contact(
  ContactArn = "string"
)
}
}

\keyword{internal}
