% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_get_profile}
\alias{alexaforbusiness_get_profile}
\title{Gets the details of a room profile by profile ARN}
\usage{
alexaforbusiness_get_profile(ProfileArn)
}
\arguments{
\item{ProfileArn}{The ARN of the room profile for which to request details. Required.}
}
\description{
Gets the details of a room profile by profile ARN.
}
\section{Request syntax}{
\preformatted{svc$get_profile(
  ProfileArn = "string"
)
}
}

\keyword{internal}
