% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_list_business_report_schedules}
\alias{alexaforbusiness_list_business_report_schedules}
\title{Lists the details of the schedules that a user configured}
\usage{
alexaforbusiness_list_business_report_schedules(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The token used to list the remaining schedules from the previous API
call.}

\item{MaxResults}{The maximum number of schedules listed in the call.}
}
\description{
Lists the details of the schedules that a user configured.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$list_business_report_schedules(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
