% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_put_conference_preference}
\alias{alexaforbusiness_put_conference_preference}
\title{Sets the conference preferences on a specific conference provider at the
account level}
\usage{
alexaforbusiness_put_conference_preference(ConferencePreference)
}
\arguments{
\item{ConferencePreference}{[required] The conference preference of a specific conference provider.}
}
\description{
Sets the conference preferences on a specific conference provider at the
account level.
}
\section{Request syntax}{
\preformatted{svc$put_conference_preference(
  ConferencePreference = list(
    DefaultConferenceProviderArn = "string"
  )
)
}
}

\keyword{internal}
