% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_search_skill_groups}
\alias{alexaforbusiness_search_skill_groups}
\title{Searches skill groups and lists the ones that meet a set of filter and
sort criteria}
\usage{
alexaforbusiness_search_skill_groups(NextToken, MaxResults, Filters,
  SortCriteria)
}
\arguments{
\item{NextToken}{An optional token returned from a prior request. Use this token for
pagination of results from this action. If this parameter is specified,
the response includes only results beyond the token, up to the value
specified by \code{MaxResults}. Required.}

\item{MaxResults}{The maximum number of results to include in the response. If more
results exist than the specified \code{MaxResults} value, a token is included
in the response so that the remaining results can be retrieved.}

\item{Filters}{The filters to use to list a specified set of skill groups. The
supported filter key is SkillGroupName.}

\item{SortCriteria}{The sort order to use in listing the specified set of skill groups. The
supported sort key is SkillGroupName.}
}
\description{
Searches skill groups and lists the ones that meet a set of filter and
sort criteria.
}
\section{Request syntax}{
\preformatted{alexaforbusiness$search_skill_groups(
  NextToken = "string",
  MaxResults = 123,
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  SortCriteria = list(
    list(
      Key = "string",
      Value = "ASC"|"DESC"
    )
  )
)
}
}

\keyword{internal}
