% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexaforbusiness_operations.R
\name{alexaforbusiness_update_contact}
\alias{alexaforbusiness_update_contact}
\title{Updates the contact details by the contact ARN}
\usage{
alexaforbusiness_update_contact(ContactArn, DisplayName, FirstName,
  LastName, PhoneNumber)
}
\arguments{
\item{ContactArn}{[required] The ARN of the contact to update.}

\item{DisplayName}{The updated display name of the contact.}

\item{FirstName}{The updated first name of the contact.}

\item{LastName}{The updated last name of the contact.}

\item{PhoneNumber}{The updated phone number of the contact.}
}
\description{
Updates the contact details by the contact ARN.
}
\section{Request syntax}{
\preformatted{svc$update_contact(
  ContactArn = "string",
  DisplayName = "string",
  FirstName = "string",
  LastName = "string",
  PhoneNumber = "string"
)
}
}

\keyword{internal}
