% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_batch_update_phone_number}
\alias{chime_batch_update_phone_number}
\title{Updates phone number product types}
\usage{
chime_batch_update_phone_number(UpdatePhoneNumberRequestItems)
}
\arguments{
\item{UpdatePhoneNumberRequestItems}{[required] The request containing the phone number IDs and product types to update.}
}
\description{
Updates phone number product types. Choose from Amazon Chime Business
Calling and Amazon Chime Voice Connector product types.
}
\section{Request syntax}{
\preformatted{svc$batch_update_phone_number(
  UpdatePhoneNumberRequestItems = list(
    list(
      PhoneNumberId = "string",
      ProductType = "BusinessCalling"|"VoiceConnector"
    )
  )
)
}
}

\keyword{internal}
