% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_delete_voice_connector}
\alias{chime_delete_voice_connector}
\title{Deletes the specified Amazon Chime Voice Connector}
\usage{
chime_delete_voice_connector(VoiceConnectorId)
}
\arguments{
\item{VoiceConnectorId}{[required] The Amazon Chime Voice Connector ID.}
}
\description{
Deletes the specified Amazon Chime Voice Connector. Any phone numbers
assigned to the Amazon Chime Voice Connector must be unassigned from it
before it can be deleted.
}
\section{Request syntax}{
\preformatted{chime$delete_voice_connector(
  VoiceConnectorId = "string"
)
}
}

\keyword{internal}
