% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_get_phone_number}
\alias{chime_get_phone_number}
\title{Retrieves details for the specified phone number ID, such as
associations, capabilities, and product type}
\usage{
chime_get_phone_number(PhoneNumberId)
}
\arguments{
\item{PhoneNumberId}{[required] The phone number ID.}
}
\description{
Retrieves details for the specified phone number ID, such as
associations, capabilities, and product type.
}
\section{Request syntax}{
\preformatted{svc$get_phone_number(
  PhoneNumberId = "string"
)
}
}

\keyword{internal}
