% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chime_operations.R
\name{chime_logout_user}
\alias{chime_logout_user}
\title{Logs out the specified user from all of the devices they are currently
logged into}
\usage{
chime_logout_user(AccountId, UserId)
}
\arguments{
\item{AccountId}{[required] The Amazon Chime account ID.}

\item{UserId}{[required] The user ID.}
}
\description{
Logs out the specified user from all of the devices they are currently
logged into.
}
\section{Request syntax}{
\preformatted{svc$logout_user(
  AccountId = "string",
  UserId = "string"
)
}
}

\keyword{internal}
