% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_create_group}
\alias{workmail_create_group}
\title{Creates a group that can be used in Amazon WorkMail by calling the
RegisterToWorkMail operation}
\usage{
workmail_create_group(OrganizationId, Name)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the group is to be created.}

\item{Name}{[required] The name of the group.}
}
\description{
Creates a group that can be used in Amazon WorkMail by calling the
RegisterToWorkMail operation.
}
\section{Request syntax}{
\preformatted{svc$create_group(
  OrganizationId = "string",
  Name = "string"
)
}
}

\keyword{internal}
