% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmail_operations.R
\name{workmail_describe_group}
\alias{workmail_describe_group}
\title{Returns the data available for the group}
\usage{
workmail_describe_group(OrganizationId, GroupId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the group exists.}

\item{GroupId}{[required] The identifier for the group to be described.}
}
\description{
Returns the data available for the group.
}
\section{Request syntax}{
\preformatted{workmail$describe_group(
  OrganizationId = "string",
  GroupId = "string"
)
}
}

\keyword{internal}
